# docs/services/gateway/api/public.openapi.yaml
openapi: 3.0.3
info:
  title: Power Knowledge Platform Gateway API
  version: 0.1.0
  license:
    name: Unspecified
    url: https://github.com/Sakayori-Iroha-168/Software_Teamwork
  description: |
    Public caller-facing API contract for the gateway service.
    All stable public and service-to-service HTTP APIs must use RESTful resource-oriented paths.
    This revision treats gateway health checks, user/session,
    knowledge-owned knowledge base, document upload, processing, content, chunk, and retrieval routes,
    document-owned report generation and report settings routes, qa-owned intelligent question answering
    routes, admin-facing runtime model/parser configuration routes, admin user
    management, personal profile, required password-change routes, and gateway-owned
    app version freshness routes as active contracts. Both admin-overview and
    admin-metrics are active gateway contracts for the administration dashboard.
...
servers:
- url: /
  description: Gateway root
security:
- bearerAuth: []
...
paths:
...
  /api/v1/report-settings:
    get:
      tags:
      - report-generation
      summary: Get report generation settings
      operationId: getReportSettings
      x-owner-service: document
      x-gateway-role: Public admin entrypoint and admin authorization. Document service owns report generation settings; AI Gateway owns model profile credentials.
      description: Returns admin-visible report generation settings. Provider base URLs and API keys remain owned by AI Gateway model profiles and are never returned here.
      responses:
        '200':
          description: Report generation settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSettingsResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - report-generation
      summary: Update report generation settings
      operationId: updateReportSettings
      x-owner-service: document
      x-gateway-role: Public admin entrypoint and admin authorization. Document service validates referenced templates and AI Gateway model profile ids.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportSettingsRequest'
      responses:
        '200':
          description: Report generation settings updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSettingsUpdateResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token issued by auth and cached by gateway through a token hash; callers must not treat it as a JWT.
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
      - code
      - message
      - requestId
      properties:
        code:
          type: string
          enum:
          - validation_error
          - unauthorized
          - forbidden
          - not_found
          - conflict
          - rate_limited
          - dependency_error
          - internal_error
          - not_implemented
          - unsupported_intent
          - unsupported_mode
        message:
          type: string
        requestId:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
...
    ReportSettingsModelConfig:
      type: object
      description: Report-generation model profile reference. Provider credentials and base URLs remain owned by AI Gateway model profiles.
      properties:
        provider:
          type: string
          description: Fixed to ai-gateway for the current gateway contract.
          example: ai-gateway
        profileId:
          type: string
          description: AI Gateway model profile id used for report generation.
        model:
          type: string
        timeoutSeconds:
          type: integer
          minimum: 1
    ReportSettingsFileDefaults:
      type: object
      properties:
        defaultFormat:
          type: string
          enum:
          - docx
        defaultNumberingMode:
          type: string
          enum:
          - global
          - by_chapter
          description: global is supported for the MVP.
        defaultStyleProfileId:
          type: string
      additionalProperties: true
    ReportSettings:
      type: object
      properties:
        llm:
          $ref: '#/components/schemas/ReportSettingsModelConfig'
        defaultTemplates:
          type: object
          additionalProperties:
            type: string
          description: Map of reportType to default reportTemplateId.
        file:
          $ref: '#/components/schemas/ReportSettingsFileDefaults'
    UpdateReportSettingsRequest:
      $ref: '#/components/schemas/ReportSettings'
    ReportSettingsResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          $ref: '#/components/schemas/ReportSettings'
        requestId:
          type: string
    ReportSettingsUpdateResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          type: object
          required:
          - updatedAt
          properties:
            updatedAt:
              type: string
              format: date-time
        requestId:
          type: string
...